+++ src/u_request.c
@@ -24,6 +24,7 @@
  */
 #include <stdlib.h>
 #include <string.h>
+#include <ctype.h>
 
 #include "u_private.h"
 #include "ulfius.h"
@@ -127,6 +128,42 @@ static int ulfius_url_format_match(const char ** splitted_url, const char ** spl
 }
 
 /**
+ * Converts a hex character to its integer value
+ */
+static char from_hex(char ch) {
+  return isdigit(ch) ? ch - '0' : tolower(ch) - 'a' + 10;
+}
+
+/**
+ * Returns a url-decoded version of str
+ * IMPORTANT: be sure to free() the returned string after use
+ * Thanks Geek Hideout!
+ * http://www.geekhideout.com/urlcode.shtml
+ */
+static char * url_decode(const char * str) {
+  if (str != NULL) {
+    char * pstr = (char*)str, * buf = malloc(strlen(str) + 1), * pbuf = buf;
+    while (* pstr) {
+      if (* pstr == '%') {
+        if (pstr[1] && pstr[2]) {
+          * pbuf++ = from_hex(pstr[1]) << 4 | from_hex(pstr[2]);
+          pstr += 2;
+        }
+      } else if (* pstr == '+') { 
+        * pbuf++ = ' ';
+      } else {
+        * pbuf++ = * pstr;
+      }
+      pstr++;
+    }
+    * pbuf = '\0';
+    return buf;
+  } else {
+    return NULL;
+  }
+}
+
+/**
  * ulfius_endpoint_match
  * return the endpoint array matching the url called with the proper http method
  * the returned array always has its last value to NULL
@@ -185,12 +222,12 @@ struct _u_endpoint ** ulfius_endpoint_match(const char * method, const char * ur
  */
 int ulfius_parse_url(const char * url, const struct _u_endpoint * endpoint, struct _u_map * map, int check_utf8) {
   char * saveptr = NULL, * cur_word = NULL, * url_cpy = NULL, * url_cpy_addr = NULL;
-  char * saveptr_format = NULL, * saveptr_prefix = NULL, * cur_word_format = NULL, * url_format_cpy = NULL, * url_format_cpy_addr = NULL;
+  char * saveptr_format = NULL, * saveptr_prefix = NULL, * cur_word_format = NULL, * url_format_cpy = NULL, * url_format_cpy_addr = NULL, * concat_url_param = NULL;
 
   if (map != NULL && endpoint != NULL) {
     url_cpy = url_cpy_addr = o_strdup(url);
     url_format_cpy = url_format_cpy_addr = o_strdup(endpoint->url_prefix);
-    cur_word = strtok_r( url_cpy, ULFIUS_URL_SEPARATOR, &saveptr );
+    cur_word = url_decode(strtok_r( url_cpy, ULFIUS_URL_SEPARATOR, &saveptr ));
     if (endpoint->url_prefix != NULL && url_format_cpy == NULL) {
       y_log_message(Y_LOG_LEVEL_ERROR, "Ulfius - Error allocating memory for url_format_cpy");
     } else if (url_format_cpy != NULL) {
@@ -198,7 +235,8 @@ int ulfius_parse_url(const char * url, const struct _u_endpoint * endpoint, stru
     }
     while (cur_word_format != NULL && cur_word != NULL) {
       // Ignoring url_prefix words
-      cur_word = strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr );
+      o_free(cur_word);
+      cur_word = url_decode(strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr ));
       cur_word_format = strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr_prefix );
     }
     o_free(url_format_cpy_addr);
@@ -213,7 +251,7 @@ int ulfius_parse_url(const char * url, const struct _u_endpoint * endpoint, stru
     while (cur_word_format != NULL && cur_word != NULL) {
       if ((cur_word_format[0] == ':' || cur_word_format[0] == '@') && (!check_utf8 || utf8_check(cur_word) == NULL)) {
         if (u_map_has_key(map, cur_word_format+1)) {
-          char * concat_url_param = msprintf("%s,%s", u_map_get(map, cur_word_format+1), cur_word);
+          concat_url_param = msprintf("%s,%s", u_map_get(map, cur_word_format+1), cur_word);
           if (concat_url_param == NULL) {
             y_log_message(Y_LOG_LEVEL_ERROR, "Ulfius - Error allocating resources for concat_url_param");
             o_free(url_cpy_addr);
@@ -233,9 +271,11 @@ int ulfius_parse_url(const char * url, const struct _u_endpoint * endpoint, stru
           }
         }
       }
-      cur_word = strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr );
+      o_free(cur_word);
+      cur_word = url_decode(strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr ));
       cur_word_format = strtok_r( NULL, ULFIUS_URL_SEPARATOR, &saveptr_format );
     }
+    o_free(cur_word);
     o_free(url_cpy_addr);
     o_free(url_format_cpy_addr);
     url_cpy_addr = NULL;
@@ -272,6 +312,7 @@ int ulfius_init_request(struct _u_request * request) {
     request->http_protocol = NULL;
     request->http_verb = NULL;
     request->http_url = NULL;
+    request->url_path = NULL;
     request->proxy = NULL;
     request->network_type = U_USE_ALL;
     request->timeout = 0L;
@@ -308,6 +349,7 @@ int ulfius_clean_request(struct _u_request * request) {
     o_free(request->http_protocol);
     o_free(request->http_verb);
     o_free(request->http_url);
+    o_free(request->url_path);
     o_free(request->proxy);
     o_free(request->auth_basic_user);
     o_free(request->auth_basic_password);
@@ -365,6 +407,7 @@ int ulfius_copy_request(struct _u_request * dest, const struct _u_request * sour
     dest->http_protocol = o_strdup(source->http_protocol);
     dest->http_verb = o_strdup(source->http_verb);
     dest->http_url = o_strdup(source->http_url);
+    dest->url_path = o_strdup(source->url_path);
     dest->proxy = o_strdup(source->proxy);
     dest->network_type = source->network_type;
     dest->check_server_certificate = source->check_server_certificate;